+++ src/main/java/com/cloudbees/jenkins/plugins/amazonecs/ECSSlave.java
@@ -26,35 +26,55 @@
 package com.cloudbees.jenkins.plugins.amazonecs;
 
 import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.text.MessageFormat;
+import java.time.Duration;
 import java.util.Collections;
-
-import javax.annotation.CheckForNull;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import javax.annotation.concurrent.GuardedBy;
-
-import hudson.model.Descriptor;
-import hudson.model.TaskListener;
-import hudson.node_monitors.ResponseTimeMonitor;
-import hudson.slaves.AbstractCloudComputer;
-import hudson.slaves.AbstractCloudSlave;
-import hudson.slaves.ComputerLauncher;
-import hudson.slaves.RetentionStrategy;
-
+import java.util.List;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import javax.annotation.CheckForNull;
+import javax.annotation.Nonnull;
+
+import com.amazonaws.services.ecs.model.AccessDeniedException;
+import com.amazonaws.services.ecs.model.ClientException;
+import com.amazonaws.services.ecs.model.ClusterNotFoundException;
+import com.amazonaws.services.ecs.model.DescribeTasksRequest;
+import com.amazonaws.services.ecs.model.DescribeTasksResult;
+import com.amazonaws.services.ecs.model.InvalidParameterException;
+import com.amazonaws.services.ecs.model.ServerException;
+import com.amazonaws.services.shield.model.InvalidOperationException;
+
+import org.jenkinsci.plugins.durabletask.executors.OnceRetentionStrategy;
+import org.jvnet.localizer.Localizable;
+import org.jvnet.localizer.ResourceBundleHolder;
+
+import hudson.model.Computer;
+import hudson.model.Descriptor;
+import hudson.model.TaskListener;
+import hudson.slaves.AbstractCloudComputer;
+import hudson.slaves.AbstractCloudSlave;
+import hudson.slaves.ComputerLauncher;
+import hudson.slaves.OfflineCause;
+
 /**
- * This slave should only handle a single task and then be shutdown.
+ * This agent should only handle a single task and then be shutdown.
  *
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
 public class ECSSlave extends AbstractCloudSlave {
 
-    private static final Logger LOGGER = Logger.getLogger(ECSCloud.class.getName());
+    private static final Logger LOGGER = Logger.getLogger(ECSSlave.class.getName());
+    private static final ResourceBundleHolder HOLDER = ResourceBundleHolder.get(Messages.class);
 
     @Nonnull
     private final ECSCloud cloud;
+    @Nonnull
+    private final ECSTaskTemplate template;
 
     /**
      * AWS Resource Name (ARN) of the ECS Cluster.
@@ -70,64 +90,10 @@ public class ECSSlave extends AbstractCloudSlave {
     @CheckForNull
     private String taskArn;
 
-    private static RetentionStrategy deleteAfterFinished = new RetentionStrategy<ECSComputer>() {
-        @Override
-        public boolean isManualLaunchAllowed(ECSComputer c) {
-            return false;
-        }
-
-        @Override
-        @GuardedBy("hudson.model.Queue.lock")
-        public long check(ECSComputer c) {
-            LOGGER.log(Level.FINE, "Checking computer: {0}", c);
-
-            AbstractCloudSlave node = c.getNode();
-
-            // If the computer is NOT idle, then it is currently running some task.
-            // In this case, we are going to tell Jenkins that it can no longer accept
-            // any new tasks, which will cause it to create a new node for any subsequent
-            // tasks.
-            if(!c.isIdle() ) {
-                LOGGER.log(Level.FINE, "Computer is not idle; setting it to no longer accept tasks.");
-                c.setAcceptingTasks( false );
-            }
-
-            // If the computer IS idle AND it is no longer accepting tasks, then it has
-            // already had a task and completed it.  In this case, we are going to terminate
-            // the node.
-            if(c.isIdle() && !c.isAcceptingTasks() && node != null) {
-                LOGGER.log(Level.FINE, "Computer is idle and not accepting tasks; terminating it.");
-                try {
-                    node.terminate();
-                } catch (InterruptedException e) {
-                    LOGGER.log(Level.WARNING, "Failed to terminate " + c.getName(), e);
-                } catch (IOException e) {
-                    LOGGER.log(Level.WARNING, "Failed to terminate " + c.getName(), e);
-                }
-            }
-
-            // If the Response Time Monitor has marked this computer as not responding, then
-            // we are going to terminate the node to free up resources.
-            if (c.getOfflineCause() instanceof ResponseTimeMonitor.Data && node != null) {
-                LOGGER.log(Level.FINE, "Computer is not responding; terminating it");
-                try {
-                    node.terminate();
-                } catch (InterruptedException e) {
-                    LOGGER.log(Level.WARNING, "Failed to terminate " + c.getName(), e);
-                } catch (IOException e) {
-                    LOGGER.log(Level.WARNING, "Failed to terminate " + c.getName(), e);
-                }
-            }
-
-            // Tell Jenkins to check again in 1 minute.
-            return 1;
-        }
-
-    };
-
-    public ECSSlave(@Nonnull ECSCloud cloud, @Nonnull String name, @Nullable String remoteFS, @Nullable String labelString, @Nonnull ComputerLauncher launcher) throws Descriptor.FormException, IOException {
-        super(name, "ECS slave", remoteFS, 1, Mode.EXCLUSIVE, labelString, launcher, deleteAfterFinished, Collections.EMPTY_LIST);
+    public ECSSlave(@Nonnull ECSCloud cloud, @Nonnull String name, ECSTaskTemplate template, @Nonnull ComputerLauncher launcher) throws Descriptor.FormException, IOException {
+        super(name, "ECS Agent", template.getRemoteFSRoot(), 1, Mode.EXCLUSIVE, template.getLabel(), launcher, new OnceRetentionStrategy(cloud.getRetentionTimeout()), Collections.emptyList());
         this.cloud = cloud;
+        this.template = template;
     }
 
     public String getClusterArn() {
@@ -140,6 +106,10 @@ public class ECSSlave extends AbstractCloudSlave {
 
     public String getTaskArn() {
         return taskArn;
+    }
+
+    public ECSTaskTemplate getTemplate() {
+        return template;
     }
 
     void setClusterArn(String clusterArn) {
@@ -155,18 +125,42 @@ public class ECSSlave extends AbstractCloudSlave {
     }
 
     @Override
-    public AbstractCloudComputer createComputer() {
+    public AbstractCloudComputer<ECSSlave> createComputer() {
         return new ECSComputer(this);
     }
 
     @Override
     protected void _terminate(TaskListener listener) throws IOException, InterruptedException {
-        if (taskArn != null) {
-            cloud.deleteTask(taskArn, clusterArn);
+        if (taskArn == null) {
+            throw new IllegalArgumentException("taskArn is null");
+        }
+        if (clusterArn == null) {
+            throw new IllegalArgumentException("clusterArn is null");
+        }
+
+        try {
+            LOGGER.log(Level.INFO, "[{0}]: Stopping: TaskArn {1}, ClusterArn {2}", new Object[]{this.getNodeName(), taskArn, clusterArn});
+
+            cloud.getEcsService().stopTask(taskArn, clusterArn);
+
+        } catch (ServerException ex) {
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Failed to stop Task {1}", this.getNodeName(), taskArn), ex);
+        } catch (AccessDeniedException ex) {
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: No permission to stop task {1}", this.getNodeName(), taskArn), ex);
+        } catch (ClientException ex) {
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Failed to stop Task {1} due {2}", this.getNodeName(), taskArn, ex.getErrorMessage()), ex);
+        } catch (InvalidParameterException ex) {
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Failed to stop Task {1}", this.getNodeName(), taskArn), ex);
+        } catch (ClusterNotFoundException ex) {
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Cluster {1} not found", this.getNodeName(), clusterArn));
         }
     }
 
-    public ECSCloud getCloud() {
-        return cloud;
+    private void writeObject(ObjectOutputStream stream) throws IOException {
+        stream.defaultWriteObject();
+    }
+
+    private void readObject(ObjectInputStream stream) throws IOException, ClassNotFoundException {
+        stream.defaultReadObject();
     }
 }