+++ Changelog.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## vNext
+- Feature - Declarative agent configuration #79
+- BUG - Fixed typo in configuration (slaveTimoutInSeconds) wasn't saved properly
+
+
+## 1.19
+- BUG - [JENKINS-54898] ECS Plugin 1.18 cannot launch agents #77
+- BUG - [JENKINS-55078] No longer able to start ECS agents #78
+
+BREAKING: Fixed typo in config: `slaveTimoutInSeconds` is now `slaveTimeoutInSeconds`
+
+
+## 1.18
+- Feature - Introduce ECSLauncher #72
+- Feature - [JENKINS-54380] Private repository authentication #74
+- Feature - [JENKINS-52032] Replace slave occurrence by agent in the Amazon ECS plugin
+- Feature - New setting for retention timeout
+- Feature - [JENKINS-54473] Rename to agents where possible
+- BUG - [JENKINS-54473] Add support for Windows containers NetworkMode, which must be null. #76
+
+## 1.17
+- Feature - Enabling use of awsvpc network mode with the EC2 launch type #68
+- Feature - Use [OnceRetentionStrategy](https://javadoc.jenkins.io/plugin/durable-task/org/jenkinsci/plugins/durabletask/executors/OnceRetentionStrategy.html) #71
+- Bug - [[JENKINS-51775](https://issues.jenkins-ci.org/browse/)] New ECS task definition created for each instance of an Agent requested
+- BUG - Agent configuration page crashes #70
+
 ## 1.16
 - Feature - Added executionRole field for fargate support #62
 - BUG - JENKINS-37597 Remove Tasks which are not responding #63